Fix handling of empty params with next_token
Results from the SSM API are returned on a best-effort basis. An API
call may return 0 items but still yield a `next_token` key, meaning
that no results are available yet but may be available in future
calls.

Previously an empty `response.parameters` would abort parameter
fetching, even if `next_token` was not nil. This patch fixes that and
adds an rspec example for it like the existing ones.